+++ src/api/vehicles.ts
@@ -4,7 +4,7 @@ import qs from 'query-string';
 import {AxiosRequestConfig} from 'axios';
 import {WS_API_BASE_URL} from '@env';
 import {GetServiceJourneyVehicles} from './types/vehicles';
-import {useSubscription, SubscriptionEventProps} from './use-websocket';
+import {useSubscription, SubscriptionEventProps} from './use-subscription';
 
 const WEBSOCKET_BASE_URL = WS_API_BASE_URL;
